backend/requirements.txt
flask
flask-sqlalchemy
flask-login
psycopg2-binary
gunicorn
werkzeug
itsdangerous>=2.1
qrcode[pil]>=7.4
